Add /pokemon/:id/moves api route with minimal info
At the moment, a ton of requests get fired for each Move client side,
even if it's a duplicate (i.e. both a levelMove and eggMove) which could
be a source of more errors, and increased resource usage of a client.

It also makes caching more difficult as each request to pokeapi gets
cached client side, and is not utilising CloudFlare.

If this gets moved to the "server" - CloudFlare can hang onto more cache,
and clients only have to hold onto one API response which can be more optimised