Use new GET endpoint from API
The new API is located closer to the Pocketbase instance (not going via
the public internet) so this should see performance improvements.

After all endpoints have been migrated across, we will move tags out of
a pokemon/:id request and into global, client side state. This will
allow us to reduce the number of API calls that need to be made to
Pocketbase. As right now, signed in users have to make an additional
API request that can't be cached - resulting in a worse user experience
for our most valuable users.