Use new /preview route to power MonOTD
Because this needs data from either a client side store, or cookies,
and the store isn't available since a +page.ts file _might_ also run
on the server, this request needs to be made server-side.

This means we need to invalidate the route if the selected game has
been changed to ensure that the correct sprite will be returned.
This could be a problem, but since Cloudflare will cache the request
because we use query params, it should only affect the first request
of the day which would usually be a bot anyway