Migrate search away from JSON files
Search is now powered by the new API!

This means, it no longer relies on difficult to maintain JSON
dumps.

Previously, a job would run once/week that would scrape all of
Pokeapi, compare every record against a "source of truth" in
a Pocketbase DB, note any differences and if there were any, it
would just commit a new JSON file right onto the main branch of
this project.

This was:
- Difficult to maintain
- Poorly architected (why compare to PB when the json file is just
  sitting there?)
  - In AWS, which I have since moved away from and had poor
    observability over.
- Dependant on a Github PAT which had long expired (Pechurant was
  still not searchable)
- Created a disconnect between names that were displayed in search
  vs when a user would click through to a Pokemon.
- A short-term hack that was never intended to last this long

Dumping massive JSON files onto the any new user to the site is also
not the best decision I've ever made and resulted in poor lighthouse
metrics. The files would be cached by the service worker, but this
is of no use to first-time visitors.

This new system still relies on a job to scrape Pokeapi, but crucially
it writes new records to a database, which is then queried from.
Because it scrapes Pokeapi, everything that is searchable will be
warmed up in the cache with a TTL of one week - when the next scrape
will occur. This should result in significantly better performance
across the site.

Knowing when this job will run, also allows coordination with the CF
cache layer, enabling us to confidently set a TTL until the next job
runtime, meaning multiple requests to the same page will not need to
talk to the backend anymore (tbd).

The API uses the same mechanisms for determining names as the actual
pages themseles, leading to greater consistency.
It also uses the same implementation for dealing with sprites, which
allows us to finally honour the "Show version specific sprites"
setting on the search results page as well.

When a new Pokemon gets published on Pokeapi, it should now maximum
a week for it to be surfaced on the frontend without relying on file
changes or access tokens.

Items and Abilities have been removed from the search results for now.
I will follow up with these - the core API is already built.
These pages currently still link through to the legacy site anyway and
saw little usage. Re-surfacing them in search should build the foundation
to migrate away from this now 4 year old legacy page.

The JSON files are not removed yet as they are still used in a few
places. But with the /meta endpoint being part of the layout, their
last main-usage has now been removed and the removal process can
begin.