Fix issue in domain store
I'm pissed off at SvelteKit for this.
The issue was that I tried to JSON parse an empty string if the user
was not signed in. This is actually a massive fucking issue because it
breaks other parts of the framework.

If a user would now change a toggle like gender, shiny, or variety
that updates a query param, the ✨ magic ✨ that causes `load` functions
to re-run will just stop working.
This was not a problem because when the path changes entirely, like
when navigating through Pokemon themselves, the page would be reloaded
entirely.

There was no kind of error message to say that a store subscription
erroring will cause this kind of issue